feat:add employee and department management

+++ database/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS departments (
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   name VARCHAR(100) UNIQUE NOT NULL,
   description TEXT,
   created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS departments (
 );
 
 CREATE TABLE IF NOT EXISTS employees (
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   employee_number VARCHAR(50) UNIQUE NOT NULL,
   full_name VARCHAR(150) NOT NULL,
   phone VARCHAR(30),
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS employees (
   emergency_contact_name VARCHAR(150),
   emergency_contact_phone VARCHAR(30),
   job_title VARCHAR(100),
-  department_id INTEGER REFERENCES departments(id),
+  department_id BIGINT REFERENCES departments(id),
   employment_date DATE,
   employment_status VARCHAR(20) NOT NULL DEFAULT 'active',
   profile_image TEXT,
@@ -29,8 +29,8 @@ CREATE INDEX IF NOT EXISTS idx_employees_department_id
   ON employees(department_id);
 
 CREATE TABLE IF NOT EXISTS users (
-  id SERIAL PRIMARY KEY,
-  employee_id INTEGER UNIQUE REFERENCES employees(id) ON DELETE CASCADE,
+  id BIGSERIAL PRIMARY KEY,
+  employee_id BIGINT UNIQUE REFERENCES employees(id) ON DELETE CASCADE,
   email VARCHAR(255) UNIQUE NOT NULL,
   password_hash TEXT NOT NULL,
   role VARCHAR(20) NOT NULL CHECK (role IN ('admin', 'employee')),
